tidies up the files. added all the menus to menu.c
