+++ src/22_TheGits/include/menu.h
@@ -3,6 +3,8 @@
 
 void memory_menu();
 void pit_menu();
+void play_music_menu();
+void start_game_menu();
 
 
 #endif
